requirements.txt
--find-links wheelhouse
django<2
xmltodict==0.6.0
Pillow==2.7.0
lxml==3.4.1